# kafka: Reduce consumer/producer allocations