Uses more efficient kgo.Hook which reduces the allocations required per event. The result is a drop of ~40% in allocations.

```
benchstat main.txt batch_metrics.txt                                                                                                                                                                             <region:us-east-1>
goos: darwin
goarch: arm64
pkg: github.com/elastic/apm-queue/v2/kafka
                    │  main.txt   │          batch_metrics.txt         │
                    │   sec/op    │    sec/op     vs base              │
Consumer/500MaxPoll   667.5µ ± 7%   444.5µ ± 22%  -33.41% (p=0.002 n=6)

                    │   main.txt   │       batch_metrics.txt       │
                    │  consumed/s  │  consumed/s   vs base         │
Consumer/500MaxPoll   385.7k ± 18%   396.3k ± 11%  ~ (p=0.937 n=6)

                    │   main.txt    │          batch_metrics.txt       │
                    │     B/op      │     B/op       vs base           │
Consumer/500MaxPoll   1288.9Ki ± 6%   687.2Ki ± 13%  -46.68% (p=0.002 n=6)

                    │   main.txt   │         batch_metrics.txt         │
                    │  allocs/op   │  allocs/op   vs base              │
Consumer/500MaxPoll   10.898k ± 5%   6.546k ± 5%  -39.93% (p=0.002 n=6)
```